pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "updatabot"
version = "0.0.13"
authors = [
  { name="Example Author", email="author@example.com" },
]
description = "Datascript library for https://updatabot.com"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "pandas==2.2.3",
  "requests==2.32.2",
  "python-dotenv==1.0.1",
  "xlrd==2.0.1",
  "openpyxl==3.1.5",
  "pydantic==2.10.6",
  "pydantic_core==2.27.2",
  "email_validator==2.2.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]

[project.urls]
Homepage = "https://github.com/updatabot/python-updatabot"
Issues = "https://github.com/updatabot/python-updatabot/issues"